improve code organisation
