+++ push_data_to_ccos/get_repo_data.py
@@ -1,6 +1,10 @@
 #!/usr/bin/env python3
 # vim: set fileencoding=utf-8:
 
+
+# Standard library
+import sys,os
+
 # Third-party
 from github import Github
 from github.GithubException import GithubException, UnknownObjectException
@@ -8,22 +12,12 @@ import emoji
 import yaml
 
 # Local
-from push_data_via_git import GITHUB_ORGANIZATION, GITHUB_TOKEN
+sys.path.append((os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
++ '/utils/'))
+from github_utils import get_cc_organization, set_up_github_client
 
 
 CC_METADATA_FILE_NAME = ".cc-metadata.yml"
-
-
-def set_up_github_client():
-    print("Setting up GitHub client...")
-    github_client = Github(GITHUB_TOKEN)
-    return github_client
-
-
-def get_cc_organization(github_client):
-    print("Getting CC's GitHub organization...")
-    cc = github_client.get_organization(GITHUB_ORGANIZATION)
-    return cc
 
 
 def get_repositories(organization):
